(Partial) WASM support.
Some refactoring was necessary because the way emscripten/wasm works: you need
to provide a function that will be called on every frame, and you need to keep
the context that that function needs somewhere.

Some additional refactoring was done because I added more stringient compiler
flags while debugging, and lthough those changes aren't strictly needed in the
end, I'll keep them.

See wasm/README for how to build and try it out.